requirements.txt
openai==1.54.3
supabase==2.9.1
pydantic==2.9.2
python-dotenv==1.0.1
mem0ai==0.1.20
numpy==1.26.4
python-dateutil==2.9.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
typing-extensions>=4.0.0
requests>=2.31.0 